+++ src/rinoh/__init__.py
@@ -49,4 +49,5 @@ for name in CORE_MODULES:
 
 
 register_template = resource._DISTRIBUTION.register_template
+register_stylesheet = resource._DISTRIBUTION.register_stylesheet
 register_typeface = resource._DISTRIBUTION.register_typeface