+++ CHANGES.rst
@@ -20,6 +20,8 @@ Release 0.5.6 (in development)
 New Features:
 
 * Equation rendering (thanks to `ziamath <https://ziamath.readthedocs.io>`_)
+* Stylesheets can now also be registered by name at runtime. This makes them
+  referencable from other style sheet files to inherit from them.
 * Sphinx < 7.2: handle superfluous space in desc_sig_space
 * Sphinx frontend: versionmodified directive support (versionadded,
   versionchanged, versionremoved, deprecated) (#440)